// src/db/decadaires-agro/DTO.ts
export interface DecadaireAgroDTO {
    NUM_POSTE: string;
    NOM_USUEL: string;
    LAT: number;
    LON: number;
    ALTI: number;
    AAAAMM: Date;
    NUM_DECADE: number;
    RR: number | null;
    CRR: number | null;
    TN: number | null;
    CTN: number | null;
    TX: number | null;
    CTX: number | null;
    FFM: number | null;
    CFFM: number | null;
    TSVM: number | null;
    CTSVM: number | null;
    INST: number | null;
    CINST: number | null;
    GLOT: number | null;
    CGLOT: number | null;
    ETP: number | null;
}